requirements.txt
aioredis==2.0.1
anyio==3.6.2
async-timeout==4.0.2
bson==0.5.10
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
fastapi==0.88.0
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.32
protobuf==3.20.3
pydantic==1.10.4
pymongo==4.3.3
python-dateutil==2.8.2
redis==4.4.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
